ci: build, fmt, clippy, test, audit, helm and docker on push and PR
Phase 0.10. The absence of this workflow is the root cause the audit
identifies: the only workflow was tag-triggered release building, so nothing
ever verified that a commit compiled or passed tests.

The test job runs twice: once on a bare checkout, proving a clean clone is
green with only a Rust toolchain, and once against PostgreSQL and MongoDB
containers. The helm job runs `helm lint` and `helm template` with ingress and
autoscaling enabled, which is precisely what would have caught the chart
shipping values for templates that do not exist. The docker job asserts the
image does not run as root.

clippy is non-blocking for now but ratcheted: a budget file records the current
warning count and CI fails if it grows. Phase 2.2 flips it to -D warnings.

Also in this commit (Phase 2.2/2.3 hygiene):
- clippy warnings 314 -> 50, via cargo clippy --fix plus hand-fixes
- all 10 deprecated AuditLogEntry::legacy_new call sites migrated to
  ::success() with real AuditAction/ResourceType values instead of string
  matching, and legacy_new deleted
- 17 unused imports removed
- workspace lint table: unsafe_code = "forbid" locks in the zero-unsafe
  property, and three clippy lints are allowed with written justification
  rather than left as noise
- in-crate backend test suites gated on cfg(test) so they never enter a
  release build

cargo test --workspace: 330 passed, 0 failed, 10 ignored.

+++ pangolin/pangolin_api/src/token_handlers.rs
@@ -58,7 +58,7 @@ pub async fn generate_token(
     let exp = now
         .checked_add_signed(chrono::Duration::hours(expires_in as i64))
         .unwrap()
-        .timestamp() as i64;
+        .timestamp();
 
     let username = payload.username.unwrap_or_else(|| "api-user".to_string());
 
@@ -477,7 +477,7 @@ pub async fn rotate_token(
     let exp = now
         .checked_add_signed(chrono::Duration::hours(expires_in))
         .unwrap()
-        .timestamp() as i64;
+        .timestamp();
 
     let token_id = Uuid::new_v4();
     let tenant_id_str = session.tenant_id.map(|t| t.to_string()).unwrap_or_default();
